+++ EssentialFeediOSTests/Helpers/UIImage+TestHelpers.swift
@@ -1,0 +1,19 @@
+//
+// Created by Rodrigo Porto.
+// Copyright © 2024 PortoCode. All Rights Reserved.
+//
+
+import UIKit
+
+extension UIImage {
+    static func make(withColor color: UIColor) -> UIImage {
+        let rect = CGRect(x: 0, y: 0, width: 1, height: 1)
+        let format = UIGraphicsImageRendererFormat()
+        format.scale = 1
+        
+        return UIGraphicsImageRenderer(size: rect.size, format: format).image { rendererContext in
+            color.setFill()
+            rendererContext.fill(rect)
+        }
+    }
+}
